split test files to avoid source-map wasm oom
